+++ dotnet/src/GameBoy/GbGsRom.cs
@@ -1,8 +1,9 @@
 using System.Text.RegularExpressions;
 using Google.Protobuf;
+using LibreShark.Hammerhead.Nintendo64;
 using LibreShark.Hammerhead.IO;
 
-namespace LibreShark.Hammerhead.Codecs;
+namespace LibreShark.Hammerhead.GameBoy;
 
 // ReSharper disable BuiltInTypeReferenceStyle
 using u8 = Byte;
