+++ dotnet/src/GameBoyColor/GbcGsV4Rom.cs
@@ -1,7 +1,8 @@
 using Google.Protobuf;
+using LibreShark.Hammerhead.Nintendo64;
 using LibreShark.Hammerhead.IO;
 
-namespace LibreShark.Hammerhead.Codecs;
+namespace LibreShark.Hammerhead.GameBoyColor;
 
 // ReSharper disable BuiltInTypeReferenceStyle
 using u8 = Byte;
