+++ dotnet/src/GameBoyColor/GbcMonsterBrainRom.cs
@@ -1,7 +1,8 @@
 using System.Text.RegularExpressions;
+using LibreShark.Hammerhead.Nintendo64;
 using LibreShark.Hammerhead.IO;
 
-namespace LibreShark.Hammerhead.Codecs;
+namespace LibreShark.Hammerhead.GameBoyColor;
 
 // ReSharper disable BuiltInTypeReferenceStyle
 using u8 = Byte;
