+++ dotnet/src/Nintendo64/N64GbHunterRle01Decoder.cs
@@ -1,6 +1,6 @@
 using LibreShark.Hammerhead.IO;
 
-namespace LibreShark.Hammerhead.N64;
+namespace LibreShark.Hammerhead.Nintendo64;
 
 // ReSharper disable BuiltInTypeReferenceStyle
 using u8 = Byte;
