+++ dotnet/src/Nintendo64/N64XpCrypter.cs
@@ -1,6 +1,6 @@
 using System.Collections.Immutable;
 
-namespace LibreShark.Hammerhead.N64;
+namespace LibreShark.Hammerhead.Nintendo64;
 
 // ReSharper disable BuiltInTypeReferenceStyle
 using u8 = Byte;
