+++ dotnet/src/Nintendo64/N64XpScrambler.cs
@@ -1,4 +1,4 @@
-namespace LibreShark.Hammerhead.N64;
+namespace LibreShark.Hammerhead.Nintendo64;
 
 // ReSharper disable BuiltInTypeReferenceStyle
 using u8 = Byte;
