+++ dotnet/src/protobuf/n64.proto
@@ -1,10 +1,9 @@
 syntax = "proto3";
 
 import "google/protobuf/timestamp.proto";
-import "common.proto";
 
 package libreshark.hammerhead.n64;
-option csharp_namespace = "LibreShark.Hammerhead.N64";
+option csharp_namespace = "LibreShark.Hammerhead.Nintendo64";
 
 enum Xp64Opcode {
   NULL_OPCODE                     = 0x00;
